pyproject.toml
[project]
name = "hn-digest"
version = "1.0.0"
description = "Multi-language weekly Hacker News digest for Telegram channels"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Far Chinberdiev" }
]
keywords = ["hackernews", "digest", "telegram", "translation"]

dependencies = [
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "trafilatura>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
hn-digest = "hn_digest.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
